# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MzansiBuilds API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStage = (typeof ProjectStage)[keyof typeof ProjectStage];

export const ProjectStage = {
  Planning: "Planning",
  Building: "Building",
  Review: "Review",
  Completed: "Completed",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  stage: ProjectStage;
  supportNeeded: string;
  ownerClerkId: string;
  ownerName: string;
  createdAt: string;
  commentCount: number;
  milestoneCount: number;
}

export interface Comment {
  id: number;
  projectId: number;
  authorClerkId: string;
  authorName: string;
  content: string;
  createdAt: string;
}

export interface Milestone {
  id: number;
  projectId: number;
  description: string;
  createdAt: string;
}

export type ProjectDetail = Project & {
  comments: Comment[];
  milestones: Milestone[];
};

export type CreateProjectBodyStage =
  (typeof CreateProjectBodyStage)[keyof typeof CreateProjectBodyStage];

export const CreateProjectBodyStage = {
  Planning: "Planning",
  Building: "Building",
  Review: "Review",
  Completed: "Completed",
} as const;

export interface CreateProjectBody {
  title: string;
  description: string;
  stage: CreateProjectBodyStage;
  supportNeeded: string;
  ownerName: string;
}

export interface CreateCommentBody {
  content: string;
  authorName: string;
}

export interface CreateMilestoneBody {
  description: string;
}

export interface StatsSummary {
  totalProjects: number;
  completedProjects: number;
  totalComments: number;
  totalMilestones: number;
}

export interface StageCount {
  stage: string;
  count: number;
}

export type ListProjectsParams = {
  stage?: ListProjectsStage;
  support?: string;
};

export type ListProjectsStage =
  (typeof ListProjectsStage)[keyof typeof ListProjectsStage];

export const ListProjectsStage = {
  Planning: "Planning",
  Building: "Building",
  Review: "Review",
  Completed: "Completed",
} as const;