# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MzansiBuilds API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all projects (public feed)
 */
export const ListProjectsQueryParams = zod.object({
  stage: zod.enum(["Planning", "Building", "Review", "Completed"]).optional(),
  support: zod.coerce.string().optional(),
});

export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  stage: zod.enum(["Planning", "Building", "Review", "Completed"]),
  supportNeeded: zod.string(),
  ownerClerkId: zod.string(),
  ownerName: zod.string(),
  createdAt: zod.coerce.date(),
  commentCount: zod.number(),
  milestoneCount: zod.number(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  stage: zod.enum(["Planning", "Building", "Review", "Completed"]),
  supportNeeded: zod.string(),
  ownerName: zod.string(),
});

/**
 * @summary Get a single project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod
  .object({
    id: zod.number(),
    title: zod.string(),
    description: zod.string(),
    stage: zod.enum(["Planning", "Building", "Review", "Completed"]),
    supportNeeded: zod.string(),
    ownerClerkId: zod.string(),
    ownerName: zod.string(),
    createdAt: zod.coerce.date(),
    commentCount: zod.number(),
    milestoneCount: zod.number(),
  })
  .and(
    zod.object({
      comments: zod.array(
        zod.object({
          id: zod.number(),
          projectId: zod.number(),
          authorClerkId: zod.string(),
          authorName: zod.string(),
          content: zod.string(),
          createdAt: zod.coerce.date(),
        }),
      ),
      milestones: zod.array(
        zod.object({
          id: zod.number(),
          projectId: zod.number(),
          description: zod.string(),
          createdAt: zod.coerce.date(),
        }),
      ),
    }),
  );

/**
 * @summary Add a milestone to a project (owner only)
 */
export const AddMilestoneParams = zod.object({
  id: zod.coerce.number(),
});

export const AddMilestoneBody = zod.object({
  description: zod.string(),
});

/**
 * @summary Add a comment to a project
 */
export const AddCommentParams = zod.object({
  id: zod.coerce.number(),
});

export const AddCommentBody = zod.object({
  content: zod.string(),
  authorName: zod.string(),
});

/**
 * @summary List all completed projects (Celebration Wall)
 */
export const ListCompletedProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  stage: zod.enum(["Planning", "Building", "Review", "Completed"]),
  supportNeeded: zod.string(),
  ownerClerkId: zod.string(),
  ownerName: zod.string(),
  createdAt: zod.coerce.date(),
  commentCount: zod.number(),
  milestoneCount: zod.number(),
});
export const ListCompletedProjectsResponse = zod.array(
  ListCompletedProjectsResponseItem,
);

/**
 * @summary Get platform-wide stats summary
 */
export const GetStatsSummaryResponse = zod.object({
  totalProjects: zod.number(),
  completedProjects: zod.number(),
  totalComments: zod.number(),
  totalMilestones: zod.number(),
});

/**
 * @summary Get project count by stage
 */
export const GetStageBreakdownResponseItem = zod.object({
  stage: zod.string(),
  count: zod.number(),
});
export const GetStageBreakdownResponse = zod.array(
  GetStageBreakdownResponseItem,
);

/**
 * @summary List projects owned by the current user
 */
export const ListMyProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  stage: zod.enum(["Planning", "Building", "Review", "Completed"]),
  supportNeeded: zod.string(),
  ownerClerkId: zod.string(),
  ownerName: zod.string(),
  createdAt: zod.coerce.date(),
  commentCount: zod.number(),
  milestoneCount: zod.number(),
});
export const ListMyProjectsResponse = zod.array(ListMyProjectsResponseItem);