# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MzansiBuilds API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  projectId: number;
  authorClerkId: string;
  authorName: string;
  content: string;
  createdAt: Date;
}