# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MzansiBuilds API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyStage } from "./createProjectBodyStage";

export interface CreateProjectBody {
  title: string;
  description: string;
  stage: CreateProjectBodyStage;
  supportNeeded: string;
  ownerName: string;
}