# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MzansiBuilds API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStage } from "./projectStage";

export interface Project {
  id: number;
  title: string;
  description: string;
  stage: ProjectStage;
  supportNeeded: string;
  ownerClerkId: string;
  ownerName: string;
  createdAt: Date;
  commentCount: number;
  milestoneCount: number;
}