# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MzansiBuilds API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from "./comment";
import type { Milestone } from "./milestone";
import type { Project } from "./project";

export type ProjectDetail = Project & {
  comments: Comment[];
  milestones: Milestone[];
};