# lib/api-zod/src/generated/types/stageCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MzansiBuilds API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StageCount {
  stage: string;
  count: number;
}